Remove ldisc.h.
It only had to be a header file because ldisc.c and ldiscucs.c had to
share the structure definition. But ldiscucs.c vanished a couple of
years ago in commit 71e42b04a57479f, so it's now fine to make the
Ldisc structure definition local to ldisc.c itself, the way it should
be.